+++ src/decorators/providers/component.js
@@ -22,9 +22,19 @@ export const Component = config => t => {
 	providerWriter.set('type', TYPE, t);
 
 	// Sensible defaults for components
-	componentWriter.set('restrict', restrict, t);
-	componentWriter.set('scope', {}, t);
-	componentWriter.set('bindToController', true, t);
+	if( !componentWriter.has('restrict', t) )
+	{
+		componentWriter.set('restrict', restrict, t);
+	}
+	if( !componentWriter.has('scope', t) )
+	{
+		componentWriter.set('scope', {}, t);
+	}
+	if( !componentWriter.has('bindToController', t) )
+	{
+		componentWriter.set('bindToController', true, t);
+	}
+
 	componentWriter.set('controllerAs', name, t);
 
 	decorateDirective(config, t);